flask_api/requirements.txt
numpy
torchvision
keras
sklearn
flask
tensorflow
